jobs/fitbit_loader/requirements.txt
fitbit==0.3.0
pymongo==3.7.2
python-dateutil==2.7.3
